+++ sonar-db/src/main/java/org/sonar/db/user/GroupMapper.java
@@ -28,7 +28,7 @@ import org.apache.ibatis.session.RowBounds;
 public interface GroupMapper {
 
   @CheckForNull
-  GroupDto selectById(long groupId);
+  GroupDto selectById(int groupId);
 
   List<GroupDto> selectByUserLogin(String userLogin);
 
@@ -47,9 +47,9 @@ public interface GroupMapper {
    *
    * @return 1 or 0. Either because the organization uuid is not the one of the group or because the group does not exist
    */
-  int countGroupByOrganizationAndId(@Param("organizationUuid") String organizationUuid, @Param("groupId") long groupId);
+  int countGroupByOrganizationAndId(@Param("organizationUuid") String organizationUuid, @Param("groupId") int groupId);
 
-  void deleteById(long groupId);
+  void deleteById(int groupId);
 
   void deleteByOrganization(@Param("organizationUuid") String organizationUuid);
 
@@ -58,5 +58,5 @@ public interface GroupMapper {
 
   List<GroupDto> selectByOrganizationUuid(@Param("organizationUuid") String organizationUuid);
 
-  List<GroupDto> selectByIds(@Param("ids") List<Long> ids);
+  List<GroupDto> selectByIds(@Param("ids") List<Integer> ids);
 }
